requirements.txt
# Vercel Serverless - Production Dependencies
grpcio>=1.50.0
grpcio-tools>=1.50.0
PyJWT>=2.8.0
flask>=2.0
numpy>=1.20
torch>=2.0
transformers>=4.30
vercel>=0.0.3
werkzeug>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
stripe>=7.0
prometheus-client>=0.19
psutil>=5.9
python-dotenv>=1.0